0.0.1 fixes: -Main Exception bind -Created Application.java -Init loggers in constructor
