+++ src/main/java/ru/zulvit/Main.java
@@ -3,12 +3,11 @@ package ru.zulvit;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import org.jetbrains.annotations.NotNull;
-import ru.zulvit.logservice.LoggerService;
 
 public class Main {
     public static void main(@NotNull String[] args) {
         final Injector injector = Guice.createInjector(new LoggerModule(args));
-        injector.getInstance(LoggerService.class).waitForInput();
+        injector.getInstance(Application.class).waitForInput();
     }
 }
