+++ src/main/java/ru/zulvit/logservice/CompositeLogService.java
@@ -1,12 +1,10 @@
 package ru.zulvit.logservice;
 
 import com.google.inject.Inject;
-import ru.zulvit.CustomFormatter;
 import org.jetbrains.annotations.NotNull;
+import ru.zulvit.CustomFormatter;
 
 import java.io.IOException;
-import java.util.NoSuchElementException;
-import java.util.Scanner;
 import java.util.logging.ConsoleHandler;
 import java.util.logging.FileHandler;
 import java.util.logging.Level;
@@ -15,60 +13,44 @@ import java.util.logging.Logger;
 public class CompositeLogService implements LoggerService {
     @NotNull
     private final Logger fileLogger = Logger.getLogger(CompositeLogService.class.getName() + "1");
-
     @NotNull
     private final Logger consoleLogger = Logger.getLogger(CompositeLogService.class.getName() + "2");
-
     @NotNull
     private final String tag;
+    private static int counterLog = 0;
+
 
     @Inject
-    public CompositeLogService(@NotNull String tag) {
+    public CompositeLogService(@NotNull String tag) throws IOException {
         this.tag = tag;
+        final FileHandler fileHandler = new FileHandler("MyLogFile.log", false);
+        final ConsoleHandler consoleHandler = new ConsoleHandler();
+
+        //region file logger settings
+        fileHandler.setFormatter(new CustomFormatter());
+        fileLogger.addHandler(fileHandler);
+        fileLogger.setUseParentHandlers(false);
+        //endregion
+
+        //region console logger settings
+        consoleHandler.setFormatter(new CustomFormatter());
+        consoleLogger.addHandler(consoleHandler);
+        consoleLogger.setUseParentHandlers(false);
+        //endregion
     }
 
     @Override
-    public void waitForInput() {
-        try (Scanner scanner = new Scanner(System.in)) {
-            int counterLog = 0;
-            //region file logger settings
-            FileHandler fileHandler = new FileHandler("MyLogFile.log", false);
-            fileHandler.setFormatter(new CustomFormatter());
-            fileLogger.addHandler(fileHandler);
-            fileLogger.setUseParentHandlers(false);
-            //endregion
+    public void log(String msg) {
+        fileLogger.log(Level.INFO, counterLog + ". " +
+                "<" + tag + ">" +
+                msg +
+                "</" + tag + ">");
+        counterLog++;
 
-            //region console logger settings
-            ConsoleHandler consoleHandler = new ConsoleHandler();
-            consoleHandler.setFormatter(new CustomFormatter());
-            consoleLogger.addHandler(consoleHandler);
-            consoleLogger.setUseParentHandlers(false);
-            //endregion
-
-            System.out.println("Waiting for new lines. Key in Ctrl+D to exit.");
-            while (true) {
-                if (scanner.hasNextLine()) {
-                    String msg = scanner.nextLine();
-                    fileLogger.log(Level.INFO, counterLog + ". " +
-                            "<" + tag + ">" +
-                            msg +
-                            "</" + tag + ">");
-                    counterLog++;
-
-                    consoleLogger.log(Level.INFO, counterLog + ". " +
-                            "<" + tag + ">" +
-                            msg +
-                            "</" + tag + ">");
-                    counterLog++;
-                } else {
-                    System.exit(0);
-                }
-            }
-        } catch (IllegalStateException | NoSuchElementException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-            throw new RuntimeException(e);
-        }
+        consoleLogger.log(Level.INFO, counterLog + ". " +
+                "<" + tag + ">" +
+                msg +
+                "</" + tag + ">");
+        counterLog++;
     }
 }
